fixed UI on ams1 by adding scroll and reducing width.  reduced variables and parameters from Long -> Int and Double -> Float where possible to accomodate 32 bit Android without polluting samples with generated 64 bit compat functions

+++ app/src/main/java/land/erikblok/busyworker/ThreadController/BusyWorkThreadController.kt
@@ -4,7 +4,7 @@ import android.content.Context
 
 
 class BusyThreadController(ctx: Context) : AbstractThreadController(ctx) {
-    fun startThreads(numThreads: Int, runtime: Long, workerId: Int = 0) {
+    fun startThreads(numThreads: Int, runtime: Int, workerId: Int = 0) {
         cleanUpThreads()
 
         for (i in 1..numThreads) {
@@ -13,8 +13,8 @@ class BusyThreadController(ctx: Context) : AbstractThreadController(ctx) {
         threadList.forEach { it.start() }
         //if runtime is 0 or negative, run until stop button is pressed.
         if (runtime > 0) {
-            wakeLock?.acquire(runtime + 1000)
-            handler.sendEmptyMessageDelayed(SUBJ_STOPTHREADS, runtime)
+            wakeLock?.acquire((runtime + 1000).toLong())
+            handler.sendEmptyMessageDelayed(SUBJ_STOPTHREADS, runtime.toLong())
         }
 
     }
