+++ app/src/main/java/land/erikblok/busyworker/ThreadController/RandomThreadController.kt
@@ -12,16 +12,16 @@ class RandomThreadController(ctx: Context) : AbstractThreadController(ctx){
     }
 
 
-    fun startThreads(timestep: Long, pauseProb: Double, runtime: Long){
+    fun startThreads(timestep: Int, pauseProb: Float, runtime: Int){
         cleanUpThreads()
         threadList.add(RandomWorker(timestep=timestep, pauseProb=pauseProb,runtime=runtime))
 
         threadList.forEach { it.start() }
         //if runtime is 0 or negative, run until stop button is pressed.
         if (runtime > 0) {
-            wakeLock?.acquire(runtime + 1000)
+            wakeLock?.acquire((runtime + 1000).toLong())
             //handler will kill the thread if it doesn't self-exit in time
-            handler.sendEmptyMessageDelayed(SUBJ_STOPTHREADS, runtime + 1000)
+            handler.sendEmptyMessageDelayed(SUBJ_STOPTHREADS, (runtime + 1000).toLong())
         }
     }
